Clarify MSRV, and rework tests accordingly
The backtrace crate/feature now requires 1.38.0, but there's not a great
way to conditionally run tests based on rust version.  So the tests that
require 1.38.0 are extracted into their own subcrate, which is not tested
when testing with 1.33.0
